+++ src/symbolize/mod.rs
@@ -45,14 +45,9 @@
 //!
 //!             let Sym {
 //!                 name, addr, offset, ..
-//!             } = sym;
+//!             } = &sym;
 //!
-//!             let path = match (sym.dir, sym.file) {
-//!                 (Some(dir), Some(file)) => Some(dir.join(file)),
-//!                 (dir, file) => dir.or_else(|| file.map(PathBuf::from)),
-//!             };
-//!
-//!             let src_loc = if let (Some(path), Some(line)) = (path, sym.line) {
+//!             let src_loc = if let (Some(path), Some(line)) = (sym.to_path(), sym.line) {
 //!                 if let Some(col) = sym.column {
 //!                     format!(" {}:{line}:{col}", path.display())
 //!                 } else {